I completed the Guess program.  I realized that Random was already contained in the java.util package so I could just use the import java.util.* line to account for both the Scanner class and the Random class.
I also realized that it could be possible that the user could pick a number that is greater than the randomly chosen value, so it may spit out a negative value.  To account for that I used the Math class and used the absolute value method to make sure the difference stayed positive.